### Features

* [x] Virtual Terminal Device
* [x] EQU Directive
* [x] Local labels
* [x] DOS File System Device
* [x] OPERATING SYSTEM
  * [x] OS-reserved shadow register bank
  * [x] OS-reserved memory space
  * [x] CD, LS, DIR, and TYPE commands
  * [x] RUN command to launch a program
* [x] Add a README.md to the z80 files directory

### Bug Fixes

* [x] Fix string whitespace highlighting error
* [x] Fix Variable Table not showing variables if not in same line as label

### Other changes

* [x] Fixed clippy warnings

+++ src/cpu/tests/instructions/jumps.rs
@@ -73,7 +73,7 @@ fn test_jr_d(#[case] starting_pc: u16, #[case] displacement: u8, #[case] expecte
     cpu.memory.borrow_mut().write(cpu.pc, JR_D_OPCODE); // JR d opcode
     cpu.memory
         .borrow_mut()
-        .write(cpu.pc.wrapping_add(1), displacement as u8);
+        .write(cpu.pc.wrapping_add(1), displacement);
     cpu.tick(); // Fetch JR d
 
     assert_eq!(
@@ -106,7 +106,7 @@ fn test_jr_cc_d(
     cpu.memory.borrow_mut().write(cpu.pc, opcode); // JR cc[y-4], d opcode
     cpu.memory
         .borrow_mut()
-        .write(cpu.pc.wrapping_add(1), displacement as u8);
+        .write(cpu.pc.wrapping_add(1), displacement);
     cpu.tick(); // Fetch JR cc[y-4], d
 
     assert_eq!(
